Split the redirection tests out of dispatch/routing_test.rb into dispatch/routing/redirection_test.rb
Left all the tests that use redirection in service of testing another feature.

Removed some requires apparently unnecessary to the tests - i.e. the tests run in file-specific way (with TEST=) without them.
